Update basic tests
Added test for the function ft_strcmp, there will be two kinds of test
the first one will compare the array to it self, so every line should be
equal, on the second test a created two differents arrays with different
data, so every test must not be equal.
The function pass all test made so far.

The function doesn't expect a value of NULL