Create ft_write function
Created the write function, this function could be done in two lines of code
as the parameter passed to it is the same we will pass to syscall, but for
returning equal values as the one from the libc I added some extra checks
to return a -1 when negatives or NULL values are passed.

The function passed all the tests.