Update bonus tests
Added tests for list_size within list_push_front tests as matter of not creating
another list to test the list_size, at the end of the tests for list_push_front
there a list with LIST_LEN + 1 size.
Also added a check for list_push_front_test for comparing it with the str, there
is a special case where the string 'Hello world!' is created as the first in the
node list, and only after that I point point_list to node, with this 'Hello world!'
string being added with an str_dup it does't not exist in the string str.